requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.0.0
Flask-Login==0.6.2
Werkzeug==2.3.6
